move top-level validation into property as well
